pyproject.toml
[project]
name = "kata-ai-meeting-assistant-clever-foxes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-cognitiveservices-speech>=1.46.0",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "openai>=2.2.0",
    "pyaudio>=0.2.14",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
]
